\name{sigmoid}
\alias{sigmoid}
\title{sigmoid function}
\description{this function is a simple sigmoid function}
\usage{
sigmoid(z)
}
\arguments{
    \item{z}{real value, to which sigmoid function is applied}
}
\details{
sigmoid(z) = 1/(1+exp(-z))
}
\value{
sigmoid function applied to the real value z.
}
\examples{
sigmoid(1)
}




